+++ index.html
@@ -237,6 +237,7 @@
   </script>
   <script src="/js/main.js">
   </script>
+  <script type="module" src="main.js"></script>
 
 </body>
 </html>
